scripts for mounting a usb

+++ .scripts/unmount_phone.sh
@@ -1,2 +1,4 @@
-fusermount -u ~/phone
-rmdir ~/phone
+if [ -d "/home/max/phone" ]; then
+    fusermount -u ~/phone
+    rmdir ~/phone
+fi
